chore(lint): resolve linting errors

+++ internal/backup/job_test.go
@@ -925,9 +925,9 @@ func TestBuildBackupJob_SecurityContext(t *testing.T) {
 		{
 			name:     "empty platform defaults to pinning IDs",
 			platform: "",
-			wantUser: ptr.To(int64(constants.UserBackup)),
-			wantGrp:  ptr.To(int64(constants.GroupBackup)),
-			wantFS:   ptr.To(int64(constants.GroupBackup)),
+			wantUser: ptr.To(constants.UserBackup),
+			wantGrp:  ptr.To(constants.GroupBackup),
+			wantFS:   ptr.To(constants.GroupBackup),
 		},
 	}
